requirement.txt
asgiref==3.4.1
cffi==1.15.0
cryptography==35.0.0
Django==3.2.9
mysql-connector-python==8.0.27
protobuf==3.19.1
pycparser==2.20
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
